double für die Multiplikation

+++ Main.java
@@ -13,7 +13,7 @@ public class Main
             summe += 1;
         }
         
-        int mult = 1;
+        double mult = 1;
         
         for(int i = 2; i <= obereGrenze; ++i)
